+++ taskify-pwa/src/nostr/EventCache.ts
@@ -1,29 +1,1 @@
-import type { NostrEvent } from "nostr-tools";
-
-/**
- * Lightweight in-memory cache to avoid re-processing duplicate events.
- * This can be expanded later to persist or shard if needed.
- */
-export class EventCache {
-  private seenIds = new Set<string>();
-  private maxSize: number;
-
-  constructor(maxSize = 2048) {
-    this.maxSize = Math.max(256, maxSize);
-  }
-
-  has(event: { id?: string } | null | undefined): boolean {
-    if (!event || typeof event.id !== "string") return false;
-    return this.seenIds.has(event.id);
-  }
-
-  add(event: NostrEvent): void {
-    if (!event?.id) return;
-    this.seenIds.add(event.id);
-    if (this.seenIds.size > this.maxSize) {
-      // Simple eviction: remove the oldest inserted id.
-      const [first] = this.seenIds;
-      if (first) this.seenIds.delete(first);
-    }
-  }
-}
+export { EventCache } from "taskify-runtime-nostr";
