+++ taskify-runtime-nostr/src/index.ts
@@ -1,2 +1,4 @@
 export * from "./boardKeys.js";
 export * from "./relayUrls.js";
+export * from "./EventCache.js";
+export * from "./PublishCoordinator.js";
